* Implement Cobra * Abstract main scraper loop logic into scraper.go
